[REFACTOR] Cleanup legacy code
This PR cleans up legacy code and reorgaizes some of the project structure.

- Removed stale interface
- Removed stale examples
- Temp remove rust as it depends on chat module that we plan to phase out
- Move embeddings to contrib(experimental)
