--- cpp/README.md
@@ -1,3 +1,0 @@
-# Use MLC-Compiled Models in C++
-
-[Documentation page](https://llm.mlc.ai/docs/deploy/cli.html)
