+++ cpp/serve/engine_actions/batch_decode.cc
@@ -7,7 +7,7 @@
 
 #include <numeric>
 
-#include "../../random.h"
+#include "../../support/random.h"
 #include "../config.h"
 #include "../model.h"
 #include "../sampler/sampler.h"
