+++ cpp/support/random.h
