+++ python/mlc_llm/cli/compile.py
@@ -1,4 +1,5 @@
 """Command line entrypoint of compilation."""
+
 import argparse
 import json
 import re
@@ -6,12 +7,12 @@ from functools import partial
 from pathlib import Path
 from typing import Union
 
-from mlc_llm.help import HELP
 from mlc_llm.interface.compile import (  # pylint: disable=redefined-builtin
     ModelConfigOverride,
     OptimizationFlags,
     compile,
 )
+from mlc_llm.interface.help import HELP
 from mlc_llm.model import MODELS
 from mlc_llm.quantization import QUANTIZATION
 from mlc_llm.support.argparse import ArgumentParser
