+++ python/mlc_llm/cli/convert_weight.py
@@ -1,10 +1,11 @@
 """Command line entrypoint of weight conversion."""
+
 import argparse
 from pathlib import Path
 from typing import Union
 
-from mlc_llm.help import HELP
 from mlc_llm.interface.convert_weight import convert_weight
+from mlc_llm.interface.help import HELP
 from mlc_llm.model import MODELS
 from mlc_llm.quantization import QUANTIZATION
 from mlc_llm.support.argparse import ArgumentParser
