+++ python/mlc_llm/cli/package.py
@@ -4,7 +4,7 @@ import os
 from pathlib import Path
 from typing import Union
 
-from mlc_llm.help import HELP
+from mlc_llm.interface.help import HELP
 from mlc_llm.interface.package import package
 from mlc_llm.support.argparse import ArgumentParser
 
@@ -23,7 +23,7 @@ def main(argv):
             raise ValueError(f"Path {str(path)} is expected to be a JSON file.")
         return path
 
-    def _parse_mlc_llm_home(path: str) -> Path:
+    def _parse_mlc_source_dir(path: str) -> Path:
         os.environ["MLC_LLM_SOURCE_DIR"] = path
         return Path(path)
 
@@ -40,10 +40,10 @@ def main(argv):
         help=HELP["config_package"] + ' (default: "%(default)s")',
     )
     parser.add_argument(
-        "--mlc-llm-home",
-        type=_parse_mlc_llm_home,
+        "--mlc-source-dir",
+        type=_parse_mlc_source_dir,
         default=os.environ.get("MLC_LLM_SOURCE_DIR", None),
-        help=HELP["mlc_llm_home"] + " (default: the $MLC_LLM_SOURCE_DIR environment variable)",
+        help=HELP["mlc_source_dir"] + " (default: the $MLC_LLM_SOURCE_DIR environment variable)",
     )
     parser.add_argument(
         "--output",
@@ -53,7 +53,7 @@ def main(argv):
         help=HELP["output_package"] + ' (default: "%(default)s")',
     )
     parsed = parser.parse_args(argv)
-    if parsed.mlc_llm_home is None:
+    if parsed.mlc_source_dir is None:
         raise ValueError(
             "MLC LLM home is not specified. "
             "Please obtain a copy of MLC LLM source code by "
@@ -62,6 +62,6 @@ def main(argv):
         )
     package(
         package_config_path=parsed.package_config,
-        mlc_llm_home=parsed.mlc_llm_home,
+        mlc_source_dir=parsed.mlc_source_dir,
         output=parsed.output,
     )
