--- rust/.gitignore
@@ -1,20 +1,0 @@
-# Generated by Cargo
-# will have compiled files and executables
-debug/
-target/
-
-# Remove Cargo.lock from gitignore if creating an executable, leave it for libraries
-# More information here https://doc.rust-lang.org/cargo/guide/cargo-toml-vs-cargo-lock.html
-Cargo.lock
-
-# Generated by Rust
-**/*.rs.bk
-/examples/pkg
-
-# MSVC Windows builds of rustc generate these, which store debugging information
-*.pdb
-
-# IDE files
-.idea/
-*.iml
-.vscode/
