--- rust/Cargo.toml
@@ -1,22 +1,0 @@
-[package]
-name = "mlc-llm"
-version = "0.1.0"
-license = "Apache-2.0"
-description = "Rust API for MLC LLM."
-homepage = "https://llm.mlc.ai/"
-readme = "README.md"
-keywords = ["rust", "mlc", "llm", "tvm", "AI"]
-authors = ["MLC Contributors"]
-repository = "https://github.com/mlc-ai/mlc-llm"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-tvm-rt = { path = "../3rdparty/tvm/rust/tvm-rt", version = "0.1.0-alpha", features = [
-    "dynamic-linking",
-] }
-tracing = "0.1.32"
-derive_builder = "0.12.0"
-serde = { version = "1.0.160", features = ["derive"] }
-serde_json = "1.0.107"
