--- rust/build.rs
@@ -1,6 +1,0 @@
-fn main() {
-    let mlc_home = env!("MLC_LLM_HOME");
-
-    println!("cargo:rustc-link-lib=dylib=mlc_llm_module");
-    println!("cargo:rustc-link-search=native={}/build", mlc_home);
-}
